Re-organize test and test utility scripts.
