+++ tests/utils/test_gentrades_utils.py
@@ -1,4 +1,4 @@
-"""Utility functions used in test scripts."""
+"""Utility functions used for testing 'GenTrades'."""
 
 from collections import deque
 from datetime import datetime
@@ -9,7 +9,6 @@ from typing import Any, TypeVar
 import pandas as pd
 
 from strat_backtest.base.gen_trades import GenTrades, RiskConfig, TradingConfig
-from strat_backtest.base.stock_trade import StockTrade
 from strat_backtest.utils.constants import (
     ClosedPositionResult,
     CompletedTrades,
@@ -23,6 +22,7 @@ from strat_backtest.utils.pos_utils import (
     get_std_field,
 )
 from strat_backtest.utils.utils import display_open_trades
+from tests.utils.test_utils import update_open_pos
 
 # Create generic type variable 'T'
 T = TypeVar("T")
@@ -33,51 +33,6 @@ class GenTradesTest(GenTrades):
 
     def gen_trades(self, df_signals: pd.DataFrame) -> tuple[pd.DataFrame, pd.DataFrame]:
         return pd.DataFrame(), pd.DataFrame()
-
-
-def get_latest_record(df_sample: pd.DataFrame) -> Record:
-    """Generate dictionary from last record in sample DataFrame."""
-
-    return df_sample.iloc[-1, :].to_dict()
-
-
-def get_date_record(df_sample: pd.DataFrame, dt: str | datetime) -> Record:
-    """Generate dictionary for record with required datetime."""
-
-    df = df_sample.copy()
-
-    if isinstance(dt, str):
-        # Convert to datetime type
-        dt = datetime.strptime(dt, "%Y-%m-%d")
-
-    if dt not in df_sample["date"].to_list():
-        raise ValueError(f"'{dt}' is an invalid date!")
-
-    return df.loc[df["date"] == dt, :].to_dict(orient="records")[0]
-
-
-def gen_record(df_sample: pd.DataFrame, **kwargs) -> Record:
-    """Generate dictionary with desired key-value pair to intialize 'Record'
-    (if provided).
-
-    Args:
-        df_sample (pd.DataFrame):
-            Sample of signals DataFrame, which contains 'exit_signal' column.
-        **kwargs (Any):
-            Columns in 'df_sample' and its corresponding value.
-
-    Returns:
-        (dict[str, str | Decimal]):
-            Selected single record in 'df_sample' converted to dictionary.
-    """
-
-    # Get latest record from sample DataFrame
-    record = get_latest_record(df_sample)
-
-    for col, value in kwargs.items():
-        record[col] = value
-
-    return record
 
 
 def gen_testgentrades_inst(
@@ -111,50 +66,6 @@ def gen_testgentrades_inst(
         setattr(gen_trades, field, attribute)
 
     return gen_trades
-
-
-def update_open_pos(
-    trade: StockTrade, exit_dt: datetime | pd.Timestamp, exit_price: float
-) -> StockTrade:
-    """Update open position with exit datetime and price."""
-
-    # Get 'exit_action' based on 'entry_action'
-    exit_action = "sell" if trade.entry_action == "buy" else "buy"
-
-    # Ensure 'exit_dt' is datetime type
-    if isinstance(exit_dt, pd.Timestamp):
-        exit_dt = exit_dt.to_pydatetime()
-
-    trade.exit_datetime = exit_dt
-    trade.exit_action = exit_action
-    trade.exit_lots = trade.entry_lots
-    trade.exit_price = convert_to_decimal(exit_price)
-
-    return trade
-
-
-def create_new_pos(
-    record: Record,
-    num_lots: int,
-    open_trades: OpenTrades = deque(),
-) -> OpenTrades:
-    """Update 'open_trades' with trade info from 'record'."""
-
-    dt = record["date"]
-    entry_signal = record["entry_signal"]
-    entry_price = record["open"]  # Assume create new position at open price
-
-    new_pos = StockTrade(
-        ticker="AAPL",
-        entry_datetime=dt,
-        entry_action=entry_signal,
-        entry_lots=convert_to_decimal(num_lots),
-        entry_price=convert_to_decimal(entry_price),
-    )
-
-    open_trades.append(new_pos)
-
-    return open_trades
 
 
 def gen_takeallexit_completed_list(
@@ -262,9 +173,6 @@ def cal_percentloss_stop_price(
     av_price = (
         sum(trade.entry_price * trade.entry_lots for trade in open_trades) / total_lots
     )
-
-    display_open_trades(open_trades)
-    print(f"{Decimal(str((181.46+172.42+198.15)/3))*(1-percent_loss)=}")
 
     stop_price = (
         av_price * (1 - percent_loss)
@@ -482,11 +390,6 @@ def cal_trailing_price(
         if entry_action == "buy"
         else first_price * (1 - trigger_trail)
     )
-
-    # print(f"record : \n\n{pformat(record, sort_dicts=False)}\n")
-    # print(f"{first_price=}")
-    # print(f"{trigger_level=}")
-    # print(f"{high=}")
 
     step_level = first_price * step if step else None
